Create Turbidity-sensor
This project focuses on measuring the turbidity of water using a TS-300B turbidity sensor interfaced with an Arduino UNO microcontroller. Turbidity, which refers to the cloudiness or haziness of a liquid, is a crucial parameter in assessing water quality. High turbidity levels can indicate the presence of suspended particles, microorganisms, or other pollutants in water, which may affect its potability and overall quality.